requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
streamlit>=1.28.0
python-dotenv>=1.0.0

# LiveKit
livekit>=0.11.0
livekit-agents>=1.2.0
livekit-plugins-openai>=1.2.0
livekit-plugins-deepgram>=1.2.0

# AI/ML
openai>=1.3.0
deepgram-sdk>=3.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# HTTP & API
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.8.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Security
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Code quality
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
pylint>=3.0.0

# Monitoring
prometheus-client>=0.19.0

# Utilities
python-multipart>=0.0.6
python-dateutil>=2.8.0
pytz>=2023.3

